requirements.txt
# Add Watson NLP library and other dependencies here
ibm-watson-developer-cloud
flask  
# Add any additional libraries for formatting or testing